+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.2] - 2020-07-01
+
 ### Added
 
 - `Job` and `Service`: supports usage metrics
@@ -286,7 +288,8 @@ First stable release.
 Please see the [Releases](https://github.com/Open-EO/openeo-vue-components/releases) for changelogs prior to v1.0.0.
 
 
-[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.2.1...HEAD
+[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.2.2...HEAD
+[2.2.2]: https://github.com/Open-EO/openeo-vue-components/compare/v2.2.1...v2.2.2
 [2.2.1]: https://github.com/Open-EO/openeo-vue-components/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/Open-EO/openeo-vue-components/compare/v2.1.2...v2.2.0
 [2.1.2]: https://github.com/Open-EO/openeo-vue-components/compare/v2.1.1...v2.1.2